renaming module configurator to state
