+++ packages/pymodaq/src/pymodaq/utils/managers/overshoot/utils.py
@@ -9,7 +9,7 @@ from pymodaq_utils.config import GlobalConfig as Config, get_set_config_dir
 from pymodaq_utils.logger import set_logger, get_module_name
 from pymodaq_utils import utils
 from pymodaq_gui.parameter.pymodaq_ptypes import registerParameterType, GroupParameter
-from pymodaq.utils.managers.configurator.configurator import Configurator
+from pymodaq.utils.managers.state.configurator import Configurator
 
 config = Config()
 logger = set_logger(get_module_name(__file__))
